requirements.txt
streamlit==1.28.1
langchain==0.0.350
langchain-openai==0.0.2
langchain-chroma==0.4.18
chromadb==0.4.18
openai==1.3.7
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
sentence-transformers==2.2.2
